refactor(web): separate adapter responsibilities
