+++ packages/web/src/adapter/stagehand-factory.ts
@@ -4,80 +4,28 @@ import {
   localBrowser,
   type ModelConfig,
   Stagehand,
-  type StagehandBrowser,
   type StagehandCreateOptions,
 } from '@browserbasehq/stagehand'
-import { z } from 'zod'
 import {
   createWebEvidenceCollector,
   installWebEvidenceScript,
-  instrumentWebEvidencePages,
 } from '../evidence/web-evidence'
 import {
   defaultWebActionTimeoutMs,
   defaultWebNavigationTimeoutMs,
-  type WebAssertionDraft,
 } from '../execution-cache/web-execution-cache'
 import { abortError } from './abort'
-import { createDirectBrowser } from './direct-browser'
 import {
   type BlockedResourceType,
   blockedResourceTypes,
   type ResolvedFidelity,
 } from './fidelity'
-import type {
-  WebAutomation,
-  WebAutomationFactory,
-  WebClientContext,
-  WebIsolationState,
-  WebScreenshotCapture,
-} from './web-adapter'
+import { createStagehandAutomation } from './stagehand-automation'
+import type { WebAutomationFactory, WebClientContext } from './web-automation'
 import { defaultModelName } from './web-options'
 
 const defaultDomSettleTimeoutMs = 3_000
 const defaultObserveTimeoutMs = 10_000
-
-const verificationSchema = z.object({
-  meetsExpectation: z.boolean(),
-  actualState: z.string(),
-})
-
-const assertionLocatorShape = {
-  selector: z.string().min(1),
-  nth: z.number().int().nonnegative().optional(),
-}
-const assertionDraftSchema = z.discriminatedUnion('kind', [
-  z.strictObject({ kind: z.literal('exists'), ...assertionLocatorShape }),
-  z.strictObject({ kind: z.literal('visible'), ...assertionLocatorShape }),
-  z.strictObject({ kind: z.literal('hidden'), ...assertionLocatorShape }),
-  z.strictObject({
-    kind: z.literal('text-equals'),
-    ...assertionLocatorShape,
-    expected: z.string(),
-  }),
-  z.strictObject({
-    kind: z.literal('text-contains'),
-    ...assertionLocatorShape,
-    expected: z.string(),
-  }),
-  z.strictObject({
-    kind: z.literal('value-equals'),
-    ...assertionLocatorShape,
-    expected: z.string(),
-  }),
-  z.strictObject({
-    kind: z.literal('count-equals'),
-    ...assertionLocatorShape,
-    expected: z.union([z.number().int().nonnegative(), z.string()]),
-  }),
-  z.strictObject({ kind: z.literal('url-equals'), expected: z.string() }),
-])
-
-type StagehandTimeouts = {
-  navigationTimeoutMs: number
-  observeTimeoutMs: number
-  actTimeoutMs: number
-}
 
 type FidelityRoute = {
   request: () => { resourceType: () => string }
@@ -100,46 +48,6 @@ type FidelityBrowserContext = {
 
 function isBlockedResourceType(value: string): value is BlockedResourceType {
   return blockedResourceTypes.includes(value as BlockedResourceType)
-}
-
-function withAbort<T>(operation: Promise<T>, signal?: AbortSignal): Promise<T> {
-  if (!signal) return operation
-  if (signal.aborted) return Promise.reject(abortError())
-
-  return new Promise<T>((resolvePromise, reject) => {
-    const onAbort = () => reject(abortError())
-    signal.addEventListener('abort', onAbort, { once: true })
-    operation.then(
-      (value) => {
-        signal.removeEventListener('abort', onAbort)
-        resolvePromise(value)
-      },
-      (error) => {
-        signal.removeEventListener('abort', onAbort)
-        reject(error)
-      },
-    )
-  })
-}
-
-async function activePage(context: BrowserContext) {
-  const page = await context.activePage()
-  if (!page) throw new Error('No active browser page')
-  return page
-}
-
-async function readStagehandIsolation(
-  context: BrowserContext,
-): Promise<WebIsolationState> {
-  const cookieCount = (await context.cookies()).length
-  const page = await context.activePage()
-  let storageKeyCount = 0
-  if (page) {
-    storageKeyCount = (await page.evaluate(
-      '(() => { try { return localStorage.length } catch { return 0 } })()',
-    )) as number
-  }
-  return { cookieCount, storageKeyCount }
 }
 
 async function applyFidelity(
@@ -176,99 +84,6 @@ async function applyFidelity(
         })()
       `)
     }
-  }
-}
-
-async function createStagehandAutomation(
-  browser: StagehandBrowser,
-  stagehand: Stagehand,
-  timeouts: StagehandTimeouts,
-  evidence: ReturnType<typeof createWebEvidenceCollector>,
-): Promise<WebAutomation> {
-  const direct = createDirectBrowser(browser.context, {
-    actionTimeoutMs: timeouts.actTimeoutMs,
-    navigationTimeoutMs: timeouts.navigationTimeoutMs,
-  })
-  const instrumentPages = async () => {
-    const pages = await browser.context.pages()
-    return instrumentWebEvidencePages(pages, evidence)
-  }
-  await instrumentPages()
-  return {
-    async navigate(url, signal) {
-      await instrumentPages()
-      await direct.navigate(url, signal)
-      await instrumentPages()
-    },
-    async observe(prompt, signal) {
-      await instrumentPages()
-      const result = await withAbort(
-        stagehand.observe(prompt, {
-          timeout: timeouts.observeTimeoutMs,
-        }),
-        signal,
-      )
-      return result.data.map((action) => ({
-        description: action.description,
-        handle: action,
-      }))
-    },
-    async act(action, signal) {
-      await instrumentPages()
-      const result = await withAbort(
-        stagehand.act(action.handle as Parameters<Stagehand['act']>[0], {
-          timeout: timeouts.actTimeoutMs,
-        }),
-        signal,
-      )
-      return {
-        success: result.data.success,
-        message: result.data.success ? undefined : result.data.message,
-      }
-    },
-    async verify(prompt, signal) {
-      await instrumentPages()
-      const result = await withAbort(
-        stagehand.extract(
-          `Verify the following condition on the current page: "${prompt}". ` +
-            'Determine if the page currently meets this expectation.',
-          verificationSchema,
-        ),
-        signal,
-      )
-      return result.data
-    },
-    async compileAssertion(prompt, signal) {
-      await instrumentPages()
-      const result = await withAbort(
-        stagehand.extract(
-          'Compile the following Scenario outcome into exactly one deterministic ' +
-            'browser assertion. Use only the requested expectation, never the ' +
-            `current page value as the expected value: "${prompt}"`,
-          assertionDraftSchema,
-        ),
-        signal,
-      )
-      return result.data as WebAssertionDraft
-    },
-    async executeInstruction(instruction, bindings, signal) {
-      await instrumentPages()
-      return direct.execute(instruction, bindings, signal)
-    },
-    async screenshot(options: WebScreenshotCapture) {
-      const page = await activePage(browser.context)
-      return new Uint8Array(
-        await page.screenshot({
-          type: options.format,
-          fullPage: options.fullPage,
-        }),
-      )
-    },
-    readIsolationState: () => readStagehandIsolation(browser.context),
-    async consumeEvidence() {
-      return evidence.collect(await instrumentPages())
-    },
-    async close() {},
   }
 }
